Ignore "links" directory
I have found it useful to use dumbass-news as an HTTP server for
static files, so that Markdown plugins for web browsers format `.md`
files. (For some reason, all the plugins seem to have trouble with
formatting Markdown files on the local filesystem.)

Conventionally, I do this by putting symbolic links to the relevant
projects in the `links` directory. But it would be inappropriate for
those links to be committed to source control.